+++ modules/group_arbitrary/outputs.tf
@@ -2,6 +2,6 @@ output "created_protection_group" {
   value = aws_shield_protection_group.arbitrary
 }
 
-output "created_protections" {
-  value = aws_shield_protection.shield
-}
+# output "created_protections" {
+#   value = aws_shield_protection.shield
+# }