+++ modules/group_arbitrary/variables.tf
@@ -28,6 +28,12 @@ variable "aggregation" {
   }
 }
 
+variable "cloudfront_arns" {
+  description = "List of CloudFront ARNs"
+  type        = list(string)
+  default     = []  # You could also populate this default list with some starter ARNs if you want
+}
+
 # variable "name_resource_arn_map" {
 #   type        = map(string)
 #   description = "A map of names to resource ARNs to be protected by the Shield Advanced group."