+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 v0.0.1 - 2023-11-01
 
-- Initial release of terraform-aws-shield
+- Initial release of AWS Shield Advanced TF modules